+++ rolling/__init__.py
@@ -4,7 +4,7 @@ from .base import RollingObject
 from .entropy import Entropy
 from .logical import All, Any
 from .minmax import Min, Max, MinHeap
-from .stats import Mean, Var, Std, Median, Mode
+from .stats import Mean, Var, Std, Median, Mode, Skew
 
 def _get_subclasses(cls):
     # https://stackoverflow.com/a/33607093/3923281
@@ -39,6 +39,7 @@ def rolling(iterable, window_size, operation='Sum', window_type='fixed', **kwarg
             - 'Mean', mean value
             - 'Median', median value
             - 'Mode', set of most common values
+            - 'Skew', unbiased skewness
             - 'Var', variance of values
             - 'Std', standard deviation of values
             - 'Entropy', Shannon entropy of window