+++ rolling/tests/test_stats.py
@@ -1,10 +1,11 @@
 from collections import Counter
+from math import sqrt
 from statistics import variance, stdev, mean as _mean, median as _median
 
 import pytest
 
 from rolling.apply import Apply
-from rolling.stats import Mean, Var, Std, Median, Mode
+from rolling.stats import Mean, Var, Std, Median, Mode, Skew
 
 def _var(seq):
     if len(seq) <= 1:
@@ -28,6 +29,26 @@ def _mode(seq):
             break
         vals.add(val)
     return vals
+
+def _skew(seq):
+    """Unbiased skewness"""
+    if len(seq) <= 2:
+        return float('nan')
+
+    N = len(seq)
+
+    # compute moments
+    A = sum(seq) / N
+    B = sum(n*n for n in seq) / N - A*A
+    C = sum(n*n*n for n in seq) / N - A*A*A - 3*A*B
+
+    if B <= 1e-14:
+        return float('nan')
+
+    R = sqrt(B)
+
+    return ((sqrt(N * (N - 1)) * C) / ((N - 2) * R * R * R))
+
 
 @pytest.mark.parametrize('array', [
     [3, 0, 1, 7, 2],
@@ -131,3 +152,18 @@ def test_rolling_mode(array, window_size, window_type):
     expected = Apply(array, window_size, operation=_mode, window_type=window_type)
     # NOTE: we copy the returned set so that it is not mutated after further iteration
     assert [set_.copy() for set_ in (got)] == list(expected)
+
+
+@pytest.mark.parametrize('array', [
+    [3, -8, 1, 7, -2, 8, 1, -7, -2, 9, 3],
+    [3.2, -8.1, 4.2, 7.7, -2.1, 0, 0, -2.1, -2.9, 2.4, 3.6],
+    [3, 0, 1, 7, 2],
+    [1],
+    [],
+])
+@pytest.mark.parametrize('window_size', [3, 4, 5, 6])
+@pytest.mark.parametrize('window_type', ['fixed', 'variable'])
+def test_rolling_skew(array, window_size, window_type):
+    got = Skew(array, window_size, window_type=window_type)
+    expected = Apply(array, window_size, operation=_skew, window_type=window_type)
+    assert pytest.approx(list(got), nan_ok=True) == list(expected)